+++ cpp/mountainsort/src/utils/compute_templates_0.cpp
@@ -51,7 +51,7 @@ Mda compute_templates_0(const DiskReadMda& X, const QVector<double>& times, cons
         counts << 0;
     for (int i = 0; i < L; i++) {
         int k = labels[i];
-        int t0 = (int)(times[i] + 0.5);
+        bigint t0 = (bigint)(times[i] + 0.5);
         if (k >= 1) {
             Mda X0;
             X.readChunk(X0, 0, t0 - Tmid, M, T);
@@ -162,7 +162,7 @@ Mda32 compute_templates_0(const DiskReadMda32& X, const QVector<double>& times, 
         counts << 0;
     for (int i = 0; i < L; i++) {
         int k = labels[i];
-        bigint t0 = (int)(times[i] + 0.5);
+        bigint t0 = (bigint)(times[i] + 0.5);
         if (k >= 1) {
             Mda32 X0;
             X.readChunk(X0, 0, t0 - Tmid, M, T);
